requirements.txt
tensorflow
numpy
scikit-learn
Pillow
Flask
opencv-python
gunicorn